graph/requirements.txt
llama-index-core>=0.12.0
llama-index-embeddings-ollama>=0.5.0
llama-index-llms-ollama>=0.5.0
llama-index-readers-obsidian>=0.1.0
